requirements.txt
langchain==0.0.292
pytest==7.4.2
python-dotenv==1.0.0
openai==0.28.0
pytest-dotenv==0.5.2
google-search-results==2.4.2
streamlit==1.27.1
watchdog
duckduckgo-search=3.9.0
pandas
dask
bokeh=3.3.0
streamlit-bokeh-events=0.1.2
streamlit_js_eval=0.1.5
geopy=2.4.0